Fixing some docstrings
Moving apprise to its own file

Trying to trim down on CC
Copied from #139
