+++ arm/ripper/utils.py
@@ -2,7 +2,6 @@
 # Collection of utility functions
 import os
 import sys
-import yaml
 import logging
 import fcntl
 import subprocess
@@ -15,6 +14,7 @@ import re
 import psutil
 
 from arm.config.config import cfg
+from arm.ripper import apprise_bulk
 from arm.ui import app, db
 import arm.models.models as m
 
@@ -50,7 +50,7 @@ def notify(job, title, body):
 
     if cfg["APPRISE"] != "":
         try:
-            apprise_notify(cfg["APPRISE"], title, body)
+            apprise_bulk.apprise_notify(cfg["APPRISE"], title, body)
             logging.debug("apprise-config: " + str(cfg["APPRISE"]))
         except Exception as e:  # noqa: E722
             logging.error("Failed sending apprise notifications. " + str(e))
@@ -135,11 +135,14 @@ def fix_job_title(job):
 
 
 def move_files(basepath, filename, job, ismainfeature=False):
-    """Move files into final media directory\n
-    basepath = path to source directory\n
-    filename = name of file to be moved\n
-    job = instance of Job class\n
-    ismainfeature = True/False"""
+    """
+    Move files into final media directory\n\n
+    :param basepath: path to source directory\n
+    :param filename: name of file to be moved\n
+    :param job: instance of Job class\n
+    :param ismainfeature: True/False
+    :return: None
+    """
     type_sub_folder = convert_job_type(job.video_type)
     videotitle = fix_job_title(job)
 
@@ -313,6 +316,12 @@ def rip_data(job, datapath, logfile):
 
 
 def set_permissions(job, directory_to_traverse):
+    """
+
+    :param job: job object
+    :param directory_to_traverse: directory to fix permissions
+    :return: Bool if fails
+    """
     if not cfg['SET_MEDIA_PERMISSIONS']:
         return False
     try:
@@ -602,664 +611,3 @@ def check_ip():
             return '127.0.0.1'
     else:
         return host
-
-
-def apprise_notify(apprise_cfg, title, body):
-    """APPRISE NOTIFICATIONS
-
-    :argument
-    apprise_cfg - The full path to the apprise.yaml file
-    title - the message title
-    body - the main body of the message
-
-    :returns
-    nothing
-    """
-    yaml_file = apprise_cfg
-    with open(yaml_file, "r") as f:
-        cfg = yaml.safe_load(f)
-
-    # boxcar
-    # boxcar://{access_key}/{secret_key}
-    if cfg['BOXCAR_KEY'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-            # A sample pushbullet notification
-            apobj.add('boxcar://' + str(cfg['BOXCAR_KEY']) + "/" + str(cfg['BOXCAR_SECRET']))
-
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending boxcar apprise notification.  Continuing processing...")
-    # discord
-    # discord://{WebhookID}/{WebhookToken}/
-    if cfg['DISCORD_WEBHOOK_ID'] != "":
-        # TODO: add userid to this and config
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-            # A sample pushbullet notification
-            apobj.add('discord://' + str(cfg['DISCORD_WEBHOOK_ID']) + "/" + str(cfg['DISCORD_TOKEN']))
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending discord apprise notification.  Continuing processing...")
-    # Faast
-    # faast://{authorizationtoken}
-    if cfg['FAAST_TOKEN'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-            # A sample pushbullet notification
-            apobj.add('faast://' + str(cfg['FAAST_TOKEN']))
-
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending faast apprise notification.  Continuing processing...")
-    # FLOCK
-    # flock://{token}/
-    if cfg['FLOCK_TOKEN'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-            # A sample pushbullet notification
-            apobj.add('flock://' + str(cfg['FLOCK_TOKEN']))
-
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending flock apprise notification.  Continuing processing...")
-    # GITTER
-    # gitter: // {token} / {room} /
-    if cfg['GITTER_TOKEN'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-            # A sample pushbullet notification
-            apobj.add('gitter://' + str(cfg['GITTER_TOKEN']) + "/" + str(cfg['GITTER_ROOM']))
-
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending gitter apprise notification.  Continuing processing...")
-    # Gotify
-    # gotify://{hostname}/{token}
-    if cfg['GOTIFY_TOKEN'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-            # A sample pushbullet notification
-            apobj.add('gotify://' + str(cfg['GOTIFY_HOST']) + "/" + str(cfg['GOTIFY_TOKEN']))
-
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending gitter apprise notification.  Continuing processing...")
-    # Growl
-    # growl://{hostname} || growl://{password}@{hostname}
-    if cfg['GROWL_HOST'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-            # Check if we have a pass, use it if we do
-            if cfg['GROWL_PASS'] != "":
-                # A sample pushbullet notification
-                apobj.add('growl://' + str(cfg['GROWL_PASS']) + "@" + str(cfg['GROWL_HOST']))
-            else:
-                # A sample pushbullet notification
-                apobj.add('growl://' + str(cfg['GROWL_HOST']))
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending growl apprise notification.  Continuing processing...")
-    # JOIN
-    # join://{apikey}/ ||  join://{apikey}/{device_id}
-    if cfg['JOIN_API'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-            # Check if we have a pass, use it if we do
-            if cfg['JOIN_DEVICE'] != "":
-                # A sample pushbullet notification
-                apobj.add('join://' + str(cfg['JOIN_API']) + "/" + str(cfg['JOIN_DEVICE']))
-            else:
-                # A sample pushbullet notification
-                apobj.add('join://' + str(cfg['JOIN_API']))
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending growl apprise notification.  Continuing processing...")
-    # Kodi
-    # kodi://{hostname}:{port} || kodi: // {userid}: {password} @ {hostname}:{port}
-    if cfg['KODI_HOST'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-            # check if we have login details, if so use them
-            if cfg['KODI_USER'] != "":
-                apobj.add('kodi://' + str(cfg['KODI_USER']) + ":" + str(cfg['KODI_PASS']) + "@" + str(
-                    cfg['KODI_HOST']) + ":" + str(cfg['KODI_PORT']))
-            else:
-                if cfg['KODI_PORT'] != "":
-                    # we need to check if they are using secure or this will fail
-                    if cfg['KODI_PORT'] == "443":
-                        apobj.add('kodis://' + str(cfg['KODI_HOST']) + ":" + str(cfg['KODI_PORT']))
-                    else:
-                        apobj.add('kodi://' + str(cfg['KODI_HOST']) + ":" + str(cfg['KODI_PORT']))
-                else:
-                    apobj.add('kodi://' + str(cfg['KODI_HOST']))
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending KODI apprise notification.  Continuing processing...")
-    # KUMULOS
-    if cfg['KUMULOS_API'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-            # A sample pushbullet notification
-            apobj.add('kumulos://' + str(cfg['KUMULOS_API']) + "/" + str(cfg['KUMULOS_SERVERKEY']))
-
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending KUMULOS apprise notification.  Continuing processing...")
-    # LEMETRIC
-    if cfg['LAMETRIC_MODE'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-
-            # find the correct mode
-            if cfg['LAMETRIC_MODE'] == "device":
-                apobj.add('lametric://' + str(cfg['LAMETRIC_API']) + "@" + str(cfg['LAMETRIC_HOST']))
-            elif cfg['LAMETRIC_MODE'] == "cloud":
-                apobj.add('lametric://' + str(cfg['LAMETRIC_APP_ID']) + "@" + str(cfg['LAMETRIC_TOKEN']))
-            else:
-                logging.error("LAMETRIC apprise LAMETRIC_MODE not set")
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending LAMETRIC apprise notification.  Continuing processing...")
-    # MAILGUN
-    if cfg['MAILGUN_DOMAIN'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-            # A sample pushbullet notification
-            apobj.add('mailgun://' + str(cfg['MAILGUN_USER']) + "@" + str(cfg['MAILGUN_DOMAIN']) + "/" + str(
-                cfg['MAILGUN_APIKEY']))
-
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending mailgun apprise notification.  Continuing processing...")
-    # MATRIX
-    if cfg['MATRIX_HOST'] != "" or cfg['MATRIX_TOKEN'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-            if cfg['MATRIX_HOST'] != "":
-                apobj.add('matrixs://' + str(cfg['MATRIX_USER']) + ":" + str(cfg['MATRIX_PASS']) + "@" + str(
-                    cfg['MATRIX_HOST']))  # + "/#general/#apprise")
-            else:
-                apobj.add('matrix://' + str(cfg['MATRIX_TOKEN']))
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending Matrix apprise notification.  Continuing processing...")
-    # Microsoft teams
-    if cfg['MSTEAMS_TOKENA'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-
-            # msteams://{tokenA}/{tokenB}/{tokenC}/
-            apobj.add('msteams://' + str(cfg['MSTEAMS_TOKENA']) + "/" + str(cfg['MSTEAMS_TOKENB']) + "/" + str(
-                cfg['MSTEAMS_TOKENC']) + "/")
-
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending Microsoft teams apprise notification.  Continuing processing...")
-    # Nextcloud
-    if cfg['NEXTCLOUD_HOST'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-
-            apobj.add(
-                'nclouds://' + str(cfg['NEXTCLOUD_ADMINUSER']) + ":" + str(cfg['NEXTCLOUD_ADMINPASS']) + "@" + str(
-                    cfg['NEXTCLOUD_HOST']) + "/" + str(cfg['NEXTCLOUD_NOTIFY_USER']))
-
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending nextcloud apprise notification.  Continuing processing...")
-    # Notica
-    if cfg['NOTICA_TOKEN'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-
-            apobj.add('notica://' + str(cfg['NOTICA_TOKEN']))
-
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending notica apprise notification.  Continuing processing...")
-    # Notifico
-    if cfg['NOTIFICO_PROJECTID'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-
-            apobj.add('notica://' + str(cfg['NOTIFICO_PROJECTID']) + "/" + str(cfg['NOTIFICO_MESSAGEHOOK']))
-
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending notifico apprise notification.  continuing  processing...")
-    # Office365
-    if cfg['OFFICE365_TENANTID'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-
-            # o365://{tenant_id}:{account_email}/{client_id}/{client_secret}/
-            # TODO: we might need to escape/encode the client_secret
-            # Replace ? with %3F and  @ with %40
-            apobj.add('o365://' + str(cfg['OFFICE365_TENANTID']) + ":" + str(cfg['OFFICE365_ACCOUNTEMAIL']) + "/" + str(
-                cfg['OFFICE365_CLIENT_ID']) + "/" + str(cfg['OFFICE365_CLIENT_SECRET']))
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending Office365 apprise notification.  continuing processing...")
-    # Popcorn
-    if cfg['POPCORN_API'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-            if cfg['POPCORN_EMAIL'] != "":
-                apobj.add('popcorn://' + str(cfg['POPCORN_API']) + "/" + str(cfg['POPCORN_EMAIL']))
-            if cfg['POPCORN_PHONENO'] != "":
-                apobj.add('popcorn://' + str(cfg['POPCORN_API']) + "/" + str(cfg['POPCORN_PHONENO']))
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending popcorn apprise notification.  Continuing processing...")
-    # PROWL
-    if cfg['PROWL_API'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-            if cfg['PROWL_PROVIDERKEY'] != "":
-                apobj.add('prowl://' + str(cfg['PROWL_API']) + "/" + str(cfg['PROWL_PROVIDERKEY']))
-            else:
-                apobj.add('prowl://' + str(cfg['PROWL_API']))
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending notifico apprise notification.  continuing  processing...")
-    # Pushjet
-    # project is dead not worth coding fully
-    if cfg['PUSHJET_HOST'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-
-            apobj.add('pjet://' + str(cfg['PUSHJET_HOST']))
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending pushjet apprise notification.  continuing  processing...")
-    # techulus push
-    if cfg['PUSH_API'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-
-            apobj.add('push://' + str(cfg['PUSH_API']))
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending techulus push apprise notification.  continuing  processing...")
-    # PUSHED
-    if cfg['PUSHED_APP_KEY'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-
-            apobj.add('pushed://' + str(cfg['PUSHED_APP_KEY']) + "/" + str(cfg['PUSHED_APP_SECRET']))
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending PUSHED apprise notification.  continuing  processing...")
-    # PUSHSAFER
-    if cfg['PUSHSAFER_KEY'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-
-            apobj.add('psafers://' + str(cfg['PUSHSAFER_KEY']))
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending pushsafer apprise notification.  continuing  processing...")
-    # ROCKETCHAT
-    # rocket://{webhook}@{hostname}/{@user}
-    if cfg['ROCKETCHAT_HOST'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-            # TODO: Add checks for webhook or default modes
-            # for now only the webhook will work
-            apobj.add('rocket://' + str(cfg['ROCKETCHAT_WEBHOOK']) + "@" + str(cfg['ROCKETCHAT_HOST']))
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending rocketchat apprise notification.  continuing  processing...")
-    # ryver
-    # ryver://{organization}/{token}/
-    if cfg['RYVER_ORG'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-            # TODO: Add checks for webhook or default modes
-            # for now only the webhook will work
-            apobj.add('ryver://' + str(cfg['RYVER_ORG']) + "/" + str(cfg['RYVER_TOKEN']) + "/")
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending RYVER apprise notification.  continuing  processing...")
-    # Sendgrid
-    # sendgrid://{apikey}:{from_email}
-    if cfg['SENDGRID_API'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-            # TODO: Add tomail
-            apobj.add('sendgrid://' + str(cfg['SENDGRID_API']) + ":" + str(cfg['SENDGRID_FROMMAIL']))
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending sendgrid apprise notification.  continuing  processing...")
-    # simplepush
-    # spush://{apikey}/
-    if cfg['SIMPLEPUSH_API'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-            apobj.add('spush://' + str(cfg['SIMPLEPUSH_API']))
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending simplepush apprise notification.  continuing  processing...")
-    # slacks
-    # slack://{tokenA}/{tokenB}/{tokenC}
-    if cfg['SLACK_TOKENA'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-            apobj.add('slack://' + str(cfg['SLACK_TOKENA']) + "/" + str(cfg['SLACK_TOKENB']) + "/" + str(
-                cfg['SLACK_TOKENC']) + "/" + str(cfg['SLACK_CHANNEL']))
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending slacks apprise notification.  continuing  processing...")
-    # SPARKPOST
-    # sparkpost://{user}@{domain}/{apikey}/ || sparkpost://{user}@{domain}/{apikey}/{email}/
-    if cfg['SPARKPOST_API'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-            apobj.add('sparkpost://' + str(cfg['SPARKPOST_USER']) + "@" + str(cfg['SPARKPOST_HOST']) + "/" + str(
-                cfg['SPARKPOST_API']) + "/" + str(cfg['SPARKPOST_EMAIL']))
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending SparkPost apprise notification.  continuing  processing...")
-    # spontit
-    # spontit://{user}@{apikey}
-    if cfg['SPONTIT_API'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-            apobj.add('spontit://' + str(cfg['SPONTIT_USER_ID']) + "@" + str(cfg['SPONTIT_API']))
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending Spontit apprise notification.  continuing  processing...")
-    # Telegram
-    # tgram://{bot_token}/{chat_id}/ || tgram://{bot_token}/
-    if cfg['TELEGRAM_BOT_TOKEN'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-            apobj.add('tgram://' + str(cfg['TELEGRAM_BOT_TOKEN']) + "/" + str(cfg['TELEGRAM_CHAT_ID']))
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending Telegram apprise notification.  continuing  processing...")
-    # Twist
-    # twist://{email}/{password} || twist://{password}:{email}
-    if cfg['TWIST_EMAIL'] != "":
-        try:
-            # Create an Apprise instance
-            # TODO: add channel var and check if its blank
-            apobj = apprise.Apprise()
-            apobj.add('twist://' + str(cfg['TWIST_EMAIL']) + "/" + str(cfg['TWIST_PASS']))
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending Twist apprise notification.  continuing  processing...")
-    # XBMC
-    # xbmc://{userid}:{password}@{hostname}:{port} ||  xbmc://{hostname}:{port}
-    if cfg['XBMC_HOST'] != "":
-        try:
-            # Create an Apprise instance
-            # TODO: add channel var and check if its blank
-            apobj = apprise.Apprise()
-            # if we get user we use the username and pass
-            if cfg['XBMC_USER'] != "":
-                apobj.add('xbmc://' + str(cfg['XBMC_USER']) + ":" + str(cfg['XBMC_PASS']) + "@" + str(
-                    cfg['XBMC_HOST']) + ":" + str(cfg['XBMC_PORT']))
-            else:
-                apobj.add('xbmc://' + str(cfg['XBMC_HOST']) + ":" + str(cfg['XBMC_PORT']))
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending XBMC apprise notification.  continuing  processing...")
-    # XMPP
-    # xmpp://{password}@{hostname}:{port} || xmpps://{userid}:{password}@{hostname}
-    if cfg['XMPP_HOST'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-            # Is the user var filled
-            if cfg['XMPP_USER'] != "":
-                # xmpps://{userid}:{password}@{hostname}
-                apobj.add(
-                    'xmpps://' + str(cfg['XMPP_USER']) + ":" + str(cfg['XMPP_PASS']) + "@" + str(cfg['XMPP_HOST']))
-            else:
-                # xmpp: // {password} @ {hostname}: {port}
-                apobj.add('xmpp://' + str(cfg['XMPP_PASS']) + "@" + str(cfg['XMPP_HOST']))
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending XMPP apprise notification.  continuing  processing...")
-    # Webex teams
-    # wxteams://{token}/
-    if cfg['WEBEX_TEAMS_TOKEN'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-
-            apobj.add('wxteams://' + str(cfg['WEBEX_TEAMS_TOKEN']))
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending Webex teams apprise notification.  continuing  processing...")
-    # Zulip
-    # zulip://{botname}@{organization}/{token}/
-    if cfg['ZILUP_CHAT_TOKEN'] != "":
-        try:
-            # Create an Apprise instance
-            apobj = apprise.Apprise()
-
-            apobj.add('zulip://' + str(cfg['ZILUP_CHAT_BOTNAME']) + "@" + str(cfg['ZILUP_CHAT_ORG']) + "/" + str(
-                cfg['ZILUP_CHAT_TOKEN']))
-            # Then notify these services any time you desire. The below would
-            # notify all of the services loaded into our Apprise object.
-            apobj.notify(
-                body,
-                title=title,
-            )
-        except Exception:  # noqa: E722
-            logging.error("Failed sending Zulip apprise notification.  continuing  processing...")
